requirements.txt
# Framework Web
streamlit>=1.29.0

# OCR e Processamento de Imagem
easyocr>=1.7.1
pillow>=10.1.0
opencv-python-headless>=4.8.0

# Análise de Dados e Visualização
pandas>=2.1.0
plotly>=5.18.0
numpy>=1.26.0

# Utilitários
python-dotenv>=1.0.0
python-dateutil>=2.8.2
regex>=2023.10.3

# QR Code (consulta NFC-e)
requests
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Exportação
openpyxl>=3.1.2

# OFX (cartão)
ofxparse>=0.21

# Google Sheets (opcional: persistência remota)
gspread>=6.1.0
google-auth>=2.25.0
